[CIR-2344] Scala 3 migration, Mockito 5.18, other deps (#125)

+++ app/uk/gov/hmrc/emailverification/repositories/VerifiedHashedEmailMongoRepository.scala
@@ -39,7 +39,7 @@ class VerifiedHashedEmailMongoRepository @Inject() (mongoComponent: MongoCompone
       domainFormat = VerifiedHashedEmail.format,
       indexes = Seq(
         IndexModel(Indexes.ascending("hashedEmail"), IndexOptions().name("emailUnique").unique(true)),
-        IndexModel(Indexes.ascending("createdAt"), IndexOptions().name("ttl").expireAfter(appConfig.verifiedEmailRepoTTLDays, TimeUnit.DAYS))
+        IndexModel(Indexes.ascending("createdAt"), IndexOptions().name("ttl").expireAfter(appConfig.verifiedEmailRepoTTLDays.toLong, TimeUnit.DAYS))
       ),
       replaceIndexes = appConfig.verifiedEmailRepoReplaceIndex
     )
